0x02-functions_nested_loops

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,16 +8,16 @@
  */
 int main(void)
 {
-	unsigned long long int first = 1;
-	unsigned long long int second = 2;
-	unsigned long long int next;
+	unsigned int first = 1;
+	unsigned int second = 2;
+	unsigned int next;
 	int i;
 
 	printf("%u, %u, ", first, second);
 
 	for (i = 3; i < 50; i++)
 	{
-		unsigned long long next = first + second;
+		unsigned next = first + second;
 		printf("%u, ", next);
 		first = second;
 		second = next;
